+++ R/translate-sql-window.R
@@ -156,9 +156,8 @@ win_rank <- function(f) {
 #' @export
 win_aggregate <- function(f) {
   force(f)
-  warned <- FALSE
   function(x, na.rm = FALSE) {
-    warned <<- check_na_rm(f, na.rm, warned)
+    check_na_rm(na.rm)
     frame <- win_current_frame()
 
     win_over(